:recycle: Refactor Code
