+++ nt-exercises/src/week01/exercise03/Main.java
@@ -1,8 +1,6 @@
-package week01.exercise03.test;
+package week01.exercise03;
 
-import week01.exercise03.classes.Palindrome;
-
-public class PalindromeTest {
+public class Main {
     public static void main(String[] args) {
         Palindrome palindrome = new Palindrome();
