+++ nt-exercises/src/week01/exercise06/Main.java
@@ -1,12 +1,10 @@
-package week01.exercise06.test;
+package week01.exercise06;
 
-import week01.exercise06.classes.PaymentChange;
-
-public class PaymentChangeTest {
+public class Main {
     public static void main(String[] args) {
         PaymentChange paymentChange = new PaymentChange();
 
-        double cost = 17.35;
+        double cost = 7.35;
         double payment = 20;
 
         paymentChange.displayChange(cost,payment);
